+++ ddsweb-api/app/Http/Controllers/PrimaryBooksController.php
@@ -121,6 +121,11 @@ class PrimaryBooksController extends Controller
         $id = $request->input('id');
         try {
             $primary_book = PrimaryBook::findOrFail($id);
+
+            if(!$primary_book->okToDelete()) {
+                return response()->json(['error' => 'Cannot be deleted: Being used'],422);
+            }
+
             $primary_book->delete();
             return response()->json([
             'deleted' => true,
@@ -130,4 +135,43 @@ class PrimaryBooksController extends Controller
             return response()->json(['error' => 'Not Found'],404);
         }
     }
+
+    public function attachSourceBook(Request $request)
+    {
+        $pb_id = $request->input('id');
+        $source_book_id = $request->input('source_book');
+        try {
+            $primary_book = PrimaryBook::findOrFail($pb_id);
+            $primary_book->source_books()->attach($source_book_id);
+            return response()->json([
+            'created' => true,
+            'data' => [
+            'id' => $pb_id,
+            'source_book' => $source_book_id
+            ]
+            ], 201);
+        } catch (ModelNotFoundException $e) {
+            return response()->json(['error' => 'Not Found'],404);
+        }
+    }
+    
+    public function removeSourceBook(Request $request)
+    {
+        $pb_id = $request->input('id');
+        $source_book_id = $request->input('source_book');
+        try {
+            $primary_book = PrimaryBook::findOrFail($pb_id);
+            $primary_book->source_books()->detach($source_book_id);
+            return response()->json([
+            'deleted' => true,
+            'data' => [
+            'id' => $pb_id,
+            'source_book' => $source_book_id
+            ]
+            ], 201);
+        } catch (ModelNotFoundException $e) {
+            return response()->json(['error' => 'Not Found'],404);
+        }
+    }
+
 }
